# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  controller.downloadJobParam:
    description: Param to send to the download job
    properties:
      meta:
        description: metadata infos to format the file
        properties:
          album:
            description: the album
            type: string
          artist:
            description: the artist
            type: string
          track:
            description: the music track
            type: string
        type: object
      url:
        description: the youtube url to use for download
        type: string
    type: object
  musicresearcher.Album:
    properties:
      imageUrl:
        type: string
      name:
        type: string
      releaseDate:
        type: string
      spotifyUrl:
        type: string
    type: object
  musicresearcher.Artist:
    properties:
      genres:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
      spotifyUrl:
        type: string
    type: object
  musicresearcher.Results:
    properties:
      albums:
        items:
          $ref: '#/definitions/musicresearcher.Album'
        type: array
      artists:
        items:
          $ref: '#/definitions/musicresearcher.Artist'
        type: array
      tracks:
        items:
          $ref: '#/definitions/musicresearcher.Track'
        type: array
    type: object
  musicresearcher.Track:
    properties:
      album:
        $ref: '#/definitions/musicresearcher.Album'
      artists:
        items:
          $ref: '#/definitions/musicresearcher.Artist'
        type: array
      durationMs:
        type: integer
      name:
        type: string
      popularity:
        type: integer
      previewUrl:
        type: string
      spotifyUrl:
        type: string
    type: object
host: api.dadard.fr
info:
  contact:
    email: florian.charpentier67@gmail.com
    name: Support
  description: |-
    This application provides a front gateway allowing you
    to interact with multiple GRPC microservices hosted
    in Google Cloud
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: No terms
  title: Gateway Front API
  version: 0.0.1
paths:
  /download/url:
    post:
      consumes:
      - application/json
      description: Execute the Youtube-DL job using Cloud Task
      parameters:
      - description: Parameters to send to job
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controller.downloadJobParam'
      responses:
        "201":
          description: Created
      summary: Download and save a new music file
  /music-researcher/search:
    get:
      consumes:
      - application/json
      description: Searchs for music in Spotify API
      parameters:
      - description: Main user query
        in: query
        name: q
        required: true
        type: string
      - description: Genre list
        in: query
        items:
          type: string
        name: genre
        required: true
        type: array
      - description: Limit result count
        in: query
        name: limit
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/musicresearcher.Results'
      summary: Music search
produces:
- application/json
schemes:
- https
swagger: "2.0"